refactor: 기본차트
