+++ README.md
@@ -72,3 +72,5 @@
 - [ ] RER
 - [ ] pdf에서 sticky시에 이상함
 - [ ] 종합현황, 처치표 차트 검색, 즐찾
+- [ ] 검색 자동완성
+- [ ] 통계 : 기간동안 입원 환자, 퇴원환자, 개고양이
